requirements.txt
alpaca-py>=0.8.2
python-dotenv>=1.0.0
openai>=1.12.0
schedule>=1.2.0
pytz>=2024.1
pandas>=2.0.0
requests>=2.31.0 